fix bug in SEGMENTS_ZYX_GRID handling in synspy-register
The Z and X axes scaling factors were being swapped or rather weren't
being swapped when they should. This would distort the aspect ratio
and reduce the quality of the computed alignment transformation.

This was due to code incompletely handling the transposition between
ZYX vertex storage and the XYZ packing needed by the PCL library.