add missing limit=1 to pair study workers
These are written naively and assume they claim one work item at a
time, but the claimable work URL was missing the `?limit=1` to ensure
this.

A better fix might be to adjust the worker to allow the multi-claim to
work by processing each claimed row, but for now this is safe at
least.