honor per-image resolution info in launcher/viewer invocations
Follow the ZYX Spacing metadata from the catalog, which should be identical
to the metadata in the OME-TIFF, but allows a little more flexibility
for "rescue" of legacy data where the TIFF metadata is incomplete.

Also, override the 3D viewer ZYX_VIEW_GRID to lower resolution when
detecting nuclei, to reduce visualization resource requirements
for this task.  Since nuclei are on the order of 10 microns, try to
reduce to approximate 2 microns per pixel resolution.